refactor dark mode toggler
